[Update] GenreControllerの実装を修正
セキュリティ面や、記法の間違い等を修正した

+++ OKN/app/Http/Controllers/GenreController.php
@@ -17,9 +17,6 @@ class GenreController extends Controller
      */
     public function index()
     {
-        //
-        //return view('genres.index', ['genres' => Genre::all()]);
-        if(!Auth::check()) return redirect('login'); //確認
         return view('genres.index', ["genres" => Auth::user()->genres()->get()]);
     }
 
@@ -30,7 +27,6 @@ class GenreController extends Controller
      */
     public function create()
     {
-        if(!Auth::check()) return redirect('login'); //確認
         return view('genres.create', ['genres' => Auth::user()->genres()->get() ]);
     }
 
@@ -59,11 +55,7 @@ class GenreController extends Controller
      */
     public function show(Genre $genre)
     {
-        //
-        if(!Auth::check()) return \App::abort(404);
-        $user = Auth::user();
-        if(! $user->genres()->where('id', '=', $genre->id)->exists())
-            return \App::abort(404);
+        if($genre->user != Auth::id()) return \App::abort(404);
         return view('genres.show', ["genre" => $genre, "childs" => $user->genres()->where('parent', $genre->id)->pluck('id')]);
     }
 
@@ -75,6 +67,7 @@ class GenreController extends Controller
      */
     public function edit(Genre $genre)
     {
+        if($genre->user != Auth::id()) return \App::abort(404);
         return view('genres.edit', ["genre" => $genre, "genres" => Auth::user()->genres()->get()]);
     }
 
@@ -88,17 +81,12 @@ class GenreController extends Controller
     public function update(Request $request, Genre $genre)
     {
         // 処理が間違ってる
+        if($genre->user != Auth::id()) return \App::abort(404);
         $genre->name = $request->name;
         $genre->memo = $request->memo;
-        if($request->filled('parent')) {
-            try{
-                $user = Auth::user();
-                $parent = $user->genres()->findOrFail($request->parent);
-                $parent->child()->attach($genre->id);
-            }catch(ModelNotFoundException $e){}
-        }
+        $genre->parent = $request->parent;
         $genre->save();
-        return redirect('genres/'.$genre->id);
+        return redirect()->route('genres.show', $genre->id);
     }
 
     /**
@@ -110,7 +98,8 @@ class GenreController extends Controller
     public function destroy(Genre $genre)
     {
         //
+        if($genre->user != Auth::id()) return \App::abort(404);
         $genre->delete();
-        return redirect('genres');
+        return redirect()->route('genres.index');
     }
 }
